fix(models): split scoped catalog preparation
