+++ src/agents/prepared-model-runtime.facts.ts
@@ -500,30 +500,6 @@ export function isPreparedModelCatalogFull(snapshot: ModelCatalogSnapshot): bool
   return fullModelCatalogSnapshots.has(snapshot);
 }
 
-/** Builds a request-scoped read-only catalog from configured and auth-candidate providers. */
-export async function prepareScopedReadOnlyModelCatalog(
-  input: PreparedModelRuntimeInput,
-  providerDiscoveryProviderIds: readonly string[],
-): Promise<ModelCatalogSnapshot> {
-  const scopedInput = input.readOnly ? input : { ...input, readOnly: true };
-  const { agentFacts, workspaceFacts } = await prepareWorkspaceBuildGroup([scopedInput], "static", {
-    providerDiscoveryProviderIds,
-  });
-  const agentFactsForInput = agentFacts[0];
-  if (!agentFactsForInput) {
-    throw new Error("scoped prepared model catalog facts are missing");
-  }
-  const catalogSource = await prepareAgentCatalogSource(
-    agentFactsForInput,
-    workspaceFacts,
-    "static",
-    false,
-  );
-  return (
-    await prepareFullCatalogFacts(agentFactsForInput, workspaceFacts, "static", catalogSource)
-  ).modelCatalog;
-}
-
 function modelCatalogEntryKey(entry: Pick<ModelCatalogEntry, "id" | "provider">): string {
   return `${normalizeProviderId(entry.provider)}\0${entry.id.trim().toLowerCase()}`;
 }
